+++ crates/api/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- kwargs to Python search build ([#591](https://github.com/stac-utils/stac-rs/pull/591))
+
 ## [0.7.0] - 2025-01-02
 
 ### Added